Fix model validations and migrations, update error validations and finish tests with swagger

+++ db/migrate/20240522010508_create_carts.rb
@@ -1,7 +1,7 @@
 class CreateCarts < ActiveRecord::Migration[6.1]
   def change
     create_table :carts do |t|
-      t.decimal :total
+      t.float :total
 
       t.timestamps
     end
